requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1; python_version < '3.9'
django-crispy-forms==1.14.0
django-tinymce==3.6.1
django==4.1.7
pillow==9.4.0
sqlparse==0.4.3
